chore(kepler): resolve PR comments
