+++ domain-db/src/db/models.rs
@@ -19,7 +19,7 @@ pub struct Object {
     pub data: String,
 }
 
-#[derive(Debug, Clone, Insertable)]
+#[derive(Debug, Insertable)]
 #[diesel(table_name = objects)]
 pub struct NewObject {
     pub created_at: SystemTime,
