+++ kepler/src/main.rs
@@ -1,6 +1,6 @@
 use anyhow::{Context, Result};
 use clap::{Parser, Subcommand};
-use domain_db::{cve_sources::nist, db, db::BATCH_SIZE};
+use domain_db::{cve_sources::nist, db, db::KEPLER_BATCH_SIZE};
 use dotenvy::dotenv;
 use env_logger::Env;
 use lazy_static::lazy_static;
@@ -138,7 +138,7 @@ pub fn import_nist(
     cve_list: Vec<nist::cve::CVE>,
 ) -> Result<usize> {
     log::info!("connected to database, importing records ...");
-    log::info!("configured 'BATCH_SIZE' {}", &*BATCH_SIZE);
+    log::info!("configured 'KEPLER_BATCH_SIZE' {}", &*KEPLER_BATCH_SIZE);
     log::info!("{} CVEs pending import", cve_list.len());
 
     let mut num_imported = 0;
@@ -148,7 +148,7 @@ pub fn import_nist(
         .collect::<Vec<db::models::NewObject>>();
 
     let inserted_object_ids = repository.insert_objects(objects_to_insert)?;
-    let mut new_cves_batch: Vec<db::models::NewCVE> = Vec::with_capacity(*BATCH_SIZE);
+    let mut new_cves_batch: Vec<db::models::NewCVE> = Vec::with_capacity(*KEPLER_BATCH_SIZE);
 
     for item in &cve_list {
         let refs = item
@@ -186,7 +186,7 @@ pub fn import_nist(
             new_cves_batch.push(new_cve);
 
             // Batch insert
-            if new_cves_batch.len() >= *BATCH_SIZE {
+            if new_cves_batch.len() >= *KEPLER_BATCH_SIZE {
                 let inserted = repository.batch_insert_cves(new_cves_batch)?;
                 num_imported += inserted;
                 if num_imported > 0 {
@@ -194,7 +194,7 @@ pub fn import_nist(
                 }
 
                 // Reset the collection for the next batch
-                new_cves_batch = Vec::with_capacity(*BATCH_SIZE);
+                new_cves_batch = Vec::with_capacity(*KEPLER_BATCH_SIZE);
             }
         }
     }
